Refactor to use JSON config files
