+++ spond/experimental/glove/openimagescfg.json
@@ -1,0 +1,13 @@
+{
+  "train": true,
+  "plot": true,
+  "co_occurrence_file": "/opt/github.com/spond/spond/experimental/openimage/co_occurrence.pt",
+  "glove_options": {
+    "words_dataset": false,
+    "embed_dim": 6,
+    "n_epochs": 2000,
+    "batch_size": 1000000,
+    "x_max": 100,
+    "alpha": 0.75
+  }
+}
